+++ modules/serverpod_auth/serverpod_auth_server/lib/src/firebase/service.dart
@@ -5,6 +5,4 @@ import 'app.dart';
 abstract class FirebaseService {
   /// The app associated with the Auth service instance.
   App get app;
-
-  Future<void> delete();
 }
