+++ modules/serverpod_auth/serverpod_auth_server/lib/src/firebase/utils/error.dart
